requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.60.1
    # via matplotlib
kiwisolver==1.4.9
    # via matplotlib
matplotlib==3.10.7
    # via -r requirements.in
numpy==2.3.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
packaging==25.0
    # via matplotlib
pdf2image==1.17.0
    # via -r requirements.in
pillow==12.0.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pdf2image
pydantic==2.12.4
    # via -r requirements.in
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
pyyaml==6.0.3
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
